# Conflicts:
#	.idea/deploymentTargetDropDown.xml
#	common/src/main/res/values/styles.xml
#	data/src/main/java/com/dida/data/api/MainAPIService.kt
#	feature/nft-detail/src/main/res/layout/fragment_detail_nft.xml

+++ data/src/main/java/com/dida/data/model/response/GetPostIdCommentsResponse.kt
@@ -3,8 +3,8 @@ package com.dida.data.model.response
 import com.google.gson.annotations.SerializedName
 
 data class GetPostIdCommentsResponse(
-    @SerializedName("commentId") val commentId: Int,
-    @SerializedName("postId") val postId: Int,
+    @SerializedName("commentId") val commentId: Long,
+    @SerializedName("postId") val postId: Long,
     @SerializedName("userId") val userId: Long,
     @SerializedName("content") val content: String,
     @SerializedName("userName") val userName: String,
