+++ data/src/main/java/com/dida/data/model/response/HotSellerResponse.kt
@@ -3,7 +3,7 @@ package com.dida.data.model.response
 import com.google.gson.annotations.SerializedName
 
 data class HotSellerResponse(
-    @SerializedName("userId") val userId: Int,
+    @SerializedName("userId") val userId: Long,
     @SerializedName("name") val name: String,
     @SerializedName("imgUrl") val imgUrl: String,
     @SerializedName("profileUrl") val profileUrl: String
