+++ data/src/main/java/com/dida/data/model/response/HotUserResponse.kt
@@ -3,7 +3,7 @@ package com.dida.data.model.response
 import com.google.gson.annotations.SerializedName
 
 data class HotUserResponse(
-    @SerializedName("userId") val userId: Int,
+    @SerializedName("userId") val userId: Long,
     @SerializedName("name") val name: String,
     @SerializedName("profileUrl") val profileUrl: String,
     @SerializedName("count") val count: String,
