+++ domain/src/main/java/com/dida/domain/model/nav/post/Comments.kt
@@ -1,8 +1,8 @@
 package com.dida.domain.model.nav.post
 
 data class Comments(
-    val commentId: Int,
-    val postId: Int,
+    val commentId: Long,
+    val postId: Long,
     val userId: Long,
     val content: String,
     val userName: String,
