+++ model/src/main/java/com/dida/model/data/response/GetMainResponse.kt
@@ -10,7 +10,7 @@ data class GetMainResponse(
 )
 
 data class HotItemResponse(
-    @SerializedName("cardId") val cardId: Int,
+    @SerializedName("cardId") val cardId: Long,
     @SerializedName("imgUrl") val imgUrl: String,
     @SerializedName("name") val name: String,
     @SerializedName("price") val price: String,
@@ -18,14 +18,14 @@ data class HotItemResponse(
 )
 
 data class HotSellerResponse(
-    @SerializedName("userId") val userId: Int,
+    @SerializedName("userId") val userId: Long,
     @SerializedName("name") val name: String,
     @SerializedName("imgUrl") val imgUrl: String,
     @SerializedName("profileUrl") val profileUrl: String
 )
 
 data class RecentCardResponse(
-    @SerializedName("cardId") val cardId: Int,
+    @SerializedName("cardId") val cardId: Long,
     @SerializedName("cardName") val cardName: String,
     @SerializedName("userName") val userName: String,
     @SerializedName("imgUrl") val imgUrl: String,
@@ -34,7 +34,7 @@ data class RecentCardResponse(
 )
 
 data class HotUserResponse(
-    @SerializedName("userId") val userId: Int,
+    @SerializedName("userId") val userId: Long,
     @SerializedName("name") val name: String,
     @SerializedName("profileUrl") val profileUrl: String,
     @SerializedName("count") val count: String,
