+++ model/src/main/java/com/dida/model/domain/CardResponse.kt
@@ -1,11 +1,11 @@
 package com.dida.model.domain
 
 data class CardResponse(
-    val cardId: Int,
+    val cardId: Long,
     val title: String,
     val description: String,
     val imgUrl: String,
-    val userId: Int,
+    val userId: Long,
     val nickname: String,
     val profileUrl: String,
     val id: String,
